[rubygems/rubygems] Refactor Checksum classes and methods to reduce code. (https://github.com/rubygems/rubygems/pull/6917)
https://github.com/rubygems/rubygems/commit/2238bdaadc

+++ lib/bundler/endpoint_specification.rb
@@ -135,7 +135,7 @@ module Bundler
           else
             raise ArgumentError, "The given checksum for #{full_name} (#{digest.inspect}) is not a valid SHA256 hexdigest nor base64digest"
           end
-          @checksum = Checksum::Single.new("sha256", digest, "API response from #{@spec_fetcher.uri}")
+          @checksum = Checksum.new("sha256", digest, "API response from #{@spec_fetcher.uri}")
         when "rubygems"
           @required_rubygems_version = Gem::Requirement.new(v)
         when "ruby"
